requirements.txt
langchain==0.1.20
langchain-aws==0.1.4
langchain-community==0.0.38
boto3==1.34.69
psycopg2-binary==2.9.9
pgvector==0.2.5
reportlab==4.1.0
pandas==2.2.1
sqlparse==0.5.0
pypdf==4.2.0
